🔧 chore: 테스트 가이드에 맞게 파일명 변경, 통합 #369
